updated the computation of ym in `get_pattern_flow_data_df()` to fix bug in lag-1 computation
the old computation assumed the index years (which are assumed to be agg_year) matched to the start month, which isn't true in water year computations. Then the arrange_at(ym) was arranging things wrong, so that the lag-1 was computed between incorrect months (January to ??? instead of January to December).

- added agg_year and cal_year into df
- new internal function `agg_year_to_cal_year()`
- updated `knnst_temporal_cor()` to use the new/correct df column
- added an argument to `get_mon_plot_stats()` because there are now inconsistencies in the names of the pattern data vs the disaggregated data

closes #27